+++ rosbag2_cpp/src/rosbag2_cpp/cache/message_cache.cpp
@@ -28,7 +28,7 @@ namespace rosbag2_cpp
 namespace cache
 {
 
-MessageCache::MessageCache(uint64_t max_buffer_size)
+MessageCache::MessageCache(size_t max_buffer_size)
 {
   primary_buffer_ = std::make_shared<MessageCacheBuffer>(max_buffer_size);
   secondary_buffer_ = std::make_shared<MessageCacheBuffer>(max_buffer_size);
